refactor(code-topology): extract viz templates and split viz_3d (#108) (#111)
* refactor(code-topology): introduce handlebars render helper and templates for index and vsa

Adds shared viz_dashboard::render helper backed by handlebars, plus a
viz_dashboard::model module hosting the VizModule / VizStats types.
Moves the index and VSA visualizations (including the no-config
placeholder) from inline format! blocks into .hbs templates loaded via
include_str!.

Behavior preserved: rendered HTML matches baseline output byte-for-byte
except for the intentional dynamic timestamp on the index page.

Refs #108.

* refactor(code-topology): move clusters and codecity viz to handlebars templates

Extracts the clusters and codecity HTML+JS bodies into
templates/clusters.hbs and templates/codecity.hbs, embedded via
include_str! and rendered through the shared viz_dashboard::render
helper. Adds embed-round-trip tests so future template edits cannot
silently drop the required MODULES / COUPLING payloads.

Behavior preserved: baseline diff is only a trailing newline.

Refs #108.

* refactor(code-topology): move force_3d viz to template and split viz_3d into model / layout / render

* Extracts the force_3d dashboard viz HTML body into
  templates/force_3d.hbs (loaded via include_str!).
* Turns viz_3d.rs into a directory module:
    - viz_3d/model.rs    typed config and scene graph
    - viz_3d/layout.rs   pure force-directed simulation
    - viz_3d/render.rs   Scene3D construction plus Three.js viewer render
    - viz_3d/templates/scene.hbs   Three.js HTML template
    - viz_3d/mod.rs      public ForceDirectedProjector and Projector impl
  Public API surface (ForceDirectedProjector, ForceDirectedConfig,
  Scene3D and friends, register) is unchanged.

Behavior preserved: baseline diff for topology-3d.html and the other
viz outputs is only the trailing newline that include_str! adds.

Refs #108.

* chore(meta): collapse late-init dep_policy to satisfy clippy needless_late_init

Pre-existing needless_late_init warning that started tripping the -D warnings
gate. Restructured the if/else chain to yield dep_policy in a single
expression. No behavior change.

* test(code-topology): add structural viz tests and bump code-topology to 0.2.3

Adds a tests/viz_structure.rs integration test that asserts every
dashboard viz emits a well-shaped HTML shell, embeds its JSON payload so
it round-trips back to a serde_json::Value equal to the caller's input,
and that the escape_json_for_html guard defeats an adversarial </script>
injection payload. Also switches the CLI viz builder to consume the
shared viz_dashboard::VizModule type (with a feature-gated local fallback
so the CLI still compiles when VZ01 is off) and bumps the code-topology
crate and standard.toml versions from 0.2.2 to 0.2.3.

Closes #108.

* refactor(code-topology): share render helper and page-shell partials

Extracts the Handlebars rendering helper into an ungated
`substandards::render_shared` module so both FD01 (viz_3d) and VZ01
(viz_dashboard) can depend on it without cross-feature-gating each
other. Fixes the FD01-only build break where viz_3d/render.rs
imported `crate::substandards::viz_dashboard::render::render_template`
while viz_dashboard was gated behind VZ01.

Also introduces two shared Handlebars partials (page_head,
three_importmap) so each viz template stops duplicating the doctype
opener and the Three.js import map. Rendered output is byte-identical
to the pre-refactor bytes for codecity, force_3d, clusters, and vsa
(verified by regenerating and diffing).

Fixes stray em dashes in viz_dashboard/templates/index.hbs and
vsa.hbs per the repo's no-em-dash rule.

Bumps code-topology to 0.2.4.

* style(code-topology): remove em dashes from codecity template

Replaces em dashes with commas/colons per repo convention.
