+++ standards/v1/APS-V1-0001-code-topology/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "apss-v1-0001-code-topology"
-version = "0.2.2"
+version = "0.2.4"
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
@@ -54,6 +54,11 @@ proc-macro2 = { version = "1", optional = true }
 
 # VZ01-dashboard (VZ01 feature)
 chrono = { version = "0.4", features = ["std"], optional = true }
+
+# Handlebars: HTML templating engine shared by VZ01 (dashboard) and FD01
+# (3D force-directed) visualizations. Kept non-optional to avoid feature-gate
+# fanout across every viz submodule.
+handlebars.workspace = true
 
 [dev-dependencies]
 tempfile = "3.14"
