+++ standards/v1/APS-V1-0001-code-topology/src/substandards/partials/page_head.hbs
@@ -1,0 +1,5 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
