+++ standards/v1/APS-V1-0001-code-topology/src/substandards/partials/three_importmap.hbs
@@ -1,0 +1,8 @@
+    <script type="importmap">
+    {
+        "imports": {
+            "three": "https://cdn.jsdelivr.net/npm/three@0.160.0/build/three.module.js",
+            "three/addons/": "https://cdn.jsdelivr.net/npm/three@0.160.0/examples/jsm/"
+        }
+    }
+    </script>
