+++ standards/v1/APS-V1-0001-code-topology/src/substandards/viz_dashboard/templates/vsa_placeholder.hbs
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+<meta charset="UTF-8">
+<title>VSA Visualization  -  No Configuration</title>
+<style>
+  body { font-family: -apple-system, BlinkMacSystemFont, sans-serif; background: #1a1a2e; color: #ccc; display: flex; justify-content: center; align-items: center; min-height: 100vh; margin: 0; }
+  .card { background: #16213e; border: 1px solid #0f3460; border-radius: 12px; padding: 48px; max-width: 560px; text-align: center; }
+  h1 { color: #e94560; font-size: 1.5em; margin-bottom: 16px; }
+  p { line-height: 1.6; margin: 8px 0; }
+  code { background: #0f3460; padding: 2px 8px; border-radius: 4px; font-size: 0.9em; }
+  pre { background: #0f3460; padding: 16px; border-radius: 8px; text-align: left; overflow-x: auto; font-size: 0.85em; margin-top: 24px; }
+</style>
+</head>
+<body>
+<div class="card">
+  <h1>No VSA Configuration Found</h1>
+  <p>The VSA (Vertical Slice Architecture) visualization requires a <code>vsa.yaml</code> file in your repository root to identify which bounded contexts to display.</p>
+  <p>Without this file, all modules would appear as vertical slices  -  which is misleading for non-VSA packages.</p>
+  <pre>
+# vsa.yaml (version 1)
+version: 1
+root: ./path/to/contexts
+language: python
+
+contexts:
+  orchestration:
+    description: "Workflow execution"
+  artifacts:
+    description: "Artifact storage"</pre>
+  <p style="margin-top: 24px; font-size: 0.9em; color: #888;">See the Event Sourcing Platform docs for the full <code>vsa.yaml</code> specification.</p>
+</div>
+</body>
+</html>
