+++ standards/v1/APS-V1-0001-code-topology/standard.toml
@@ -4,7 +4,7 @@ schema = "aps.standard/v1"
 id = "APS-V1-0001"
 name = "Code Topology and Coupling Analysis"
 slug = "code-topology"
-version = "0.2.2"
+version = "0.2.4"
 category = "technical"
 status = "active"
